read paths from configuration file
This updates the automatic path detection to use a configuration file
_optionally_ as a means of overriding the paths used by
`comfyui-browser`.  This change is motivated by recent work I have done
with ComfyUI and Nix, wherein I provided [an easy means to install
`comfyui-browser` via
Nix](https://github.com/loganbarnett/dotfiles/blob/ccbfe12e755ab3d6d24b3f0f5a3386208b23357f/nix/hacks/comfyui/custom-nodes.nix#L50-L74).
The big reason this was necessary for my Nix work is because Nix's
installations are locked down into a purely read-only store.  I needed a
place to put things such as the `collections` path into a location that
`comfyui-browser` could write to.

I don't expect `comfyui-browser` to directly support Nix. My hope is
that allowing the directories to be configurable, with the defaults
being exactly what was used to before, to be something desirable for the
project in a general sense, and not just for Nix.

I've tested this under a Nix configuration with `config.json` values
being used.  Unfortunately I don't have a means to test
`comfyui-browser` outside of Nix, as the `comfyui` and CUDA setup was
hard enough with Nix itself.  So to be clear - this isn't tested under
the standard configuration that most `comfyui-browser` users are using.
I suggest giving these changes a quick spin to make sure I didn't miss
anything.  I did my best to make the code changes branchless - in that
no new code paths are added - just a default means of determining what
the configuration values should be.  Memoization is used to ensure that
the order in which these functions are called is irrelevant.

Unfortunately I am not familiar at all with Chinese/Mandarin(?) to write
the documentation for it - perhaps someone else can pick that up?

Earmarked as part of this:  I've changed the `mkdir` calls to
`makedirs`, which adds all directories in the chain that are missing,
and the `exist_ok` argument means the call will not fail if the
directory already exists, which allows us to clean up a little code.

Thanks for your work on maintaining this useful custom nodes project!